+++ android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -18,6 +18,7 @@ import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableMapKeySetIterator;
+import com.facebook.react.bridge.UiThreadUtil;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeArray;
@@ -33,8 +34,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 import io.sentry.Breadcrumb;
+import io.sentry.DateUtils;
 import io.sentry.HubAdapter;
 import io.sentry.ILogger;
 import io.sentry.ISerializer;
@@ -80,6 +84,8 @@ public class RNSentryModuleImpl {
     private static final int FROZEN_FRAME_THRESHOLD = 700;
     // 16ms (slower than 60fps) to constitute slow frames.
     private static final int SLOW_FRAME_THRESHOLD = 16;
+
+    private static final int SCREENSHOT_TIMEOUT_SECONDS = 2;
 
     public RNSentryModuleImpl(ReactApplicationContext reactApplicationContext) {
         packageInfo = getPackageInfo(reactApplicationContext);
@@ -239,11 +245,11 @@ public class RNSentryModuleImpl {
             logger.log(SentryLevel.WARNING, "App start won't be sent due to missing isColdStart.");
             promise.resolve(null);
         } else {
-            final double appStartTimestamp = appStartTime.nanoTimestamp() / 1e6;
+            final double appStartTimestampMs = DateUtils.nanosToMillis(appStartTime.nanoTimestamp());
 
             WritableMap appStart = Arguments.createMap();
 
-            appStart.putDouble("appStartTime", appStartTimestamp);
+            appStart.putDouble("appStartTime", appStartTimestampMs);
             appStart.putBoolean("isColdStart", isColdStart);
             appStart.putBoolean("didFetchAppStart", didFetchAppStart);
 
@@ -340,7 +346,8 @@ public class RNSentryModuleImpl {
             return;
         }
 
-        final byte[] raw = takeScreenshot(activity, logger, buildInfo);
+        final byte[] raw = takeScreenshotOnUiThread(activity);
+
         if (raw == null) {
             logger.log(SentryLevel.WARNING, "Screenshot is null, screen was not captured.");
             promise.resolve(null);
@@ -359,6 +366,30 @@ public class RNSentryModuleImpl {
         final WritableArray screenshotsArray = new WritableNativeArray();
         screenshotsArray.pushMap(screenshot);
         promise.resolve(screenshotsArray);
+    }
+
+    private static byte[] takeScreenshotOnUiThread(Activity activity) {
+        CountDownLatch doneSignal = new CountDownLatch(1);
+        final byte[][] bytesWrapper = {{}}; // wrapper to be able to set the value in the runnable
+        final Runnable runTakeScreenshot = () -> {
+            bytesWrapper[0] = takeScreenshot(activity, logger, buildInfo);
+            doneSignal.countDown();
+        };
+
+        if (UiThreadUtil.isOnUiThread()) {
+            runTakeScreenshot.run();
+        } else {
+            UiThreadUtil.runOnUiThread(runTakeScreenshot);
+        }
+
+        try {
+            doneSignal.await(SCREENSHOT_TIMEOUT_SECONDS, TimeUnit.SECONDS);
+        } catch (InterruptedException e) {
+            logger.log(SentryLevel.ERROR, "Screenshot process was interrupted.");
+            return null;
+        }
+
+        return bytesWrapper[0];
     }
 
     public void fetchViewHierarchy(Promise promise) {