+++ sample/src/App.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react';
-import {Provider} from 'react-redux';
+import { Provider } from 'react-redux';
 import {
   NavigationContainer,
   NavigationContainerRef,
 } from '@react-navigation/native';
-import {createStackNavigator} from '@react-navigation/stack';
+import { createStackNavigator } from '@react-navigation/stack';
 
 // Import the Sentry React Native SDK
 import * as Sentry from '@sentry/react-native';
+import * as Integrations from '@sentry/integrations';
 
 import HomeScreen from './screens/HomeScreen';
 import TrackerScreen from './screens/TrackerScreen';
@@ -15,8 +16,8 @@ import ManualTrackerScreen from './screens/ManualTrackerScreen';
 import PerformanceTimingScreen from './screens/PerformanceTimingScreen';
 import ReduxScreen from './screens/ReduxScreen';
 
-import {store} from './reduxApp';
-import {SENTRY_INTERNAL_DSN} from './dsn';
+import { store } from './reduxApp';
+import { SENTRY_INTERNAL_DSN } from './dsn';
 
 const reactNavigationInstrumentation =
   new Sentry.ReactNavigationInstrumentation({
@@ -31,7 +32,7 @@ Sentry.init({
     return e;
   },
   // This will be called with a boolean `didCallNativeInit` when the native SDK has been contacted.
-  onReady: ({didCallNativeInit}) => {
+  onReady: ({ didCallNativeInit }) => {
     console.log('onReady called with didCallNativeInit:', didCallNativeInit);
   },
   maxCacheItems: 40, // Extend from the default 30.
@@ -56,6 +57,15 @@ Sentry.init({
         console.log('Dedupe Dummy Integration');
       },
     },
+    new Integrations.HttpClient({
+      // This array can contain tuples of `[begin, end]` (both inclusive),
+      // Single status codes, or a combinations of both.
+      // default: [[500, 599]]
+      failedRequestStatusCodes: [[400, 599]],
+      // This array can contain Regexes or strings, or combinations of both.
+      // default: [/.*/]
+      failedRequestTargets: [/.*/],
+    }),
   ],
   enableAutoSessionTracking: true,
   // For testing, session close when 5 seconds (instead of the default 30) in the background.